pyproject.toml
[project]
name = "securitybot"
description = "A Discord bot that protects server security"
version = "0.1.0"
dependencies = [
    "discord.py",
    "python-dotenv"
]
requires-python = ">=3.14.2"
readme = "README.md"
keywords = ["discord", "bot", "security"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.14"
]

[[project.authors]]
name = "Kai Tomotake"
email = "kaitomotake@gmail.com"

[project.urls]
Repository = "https://github.com/Ent3r-Dev/SecurityBOT"
